move Browser test fixtures to tests/browser/conftest.py
